requirements.txt
Flask==3.0.3
gunicorn==21.2.0
textblob==0.18.0.post0
numpy==1.26.4
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
transformers==4.41.2
pillow==10.3.0
